Make corrections in 0x10-9

+++ 0x10-variadic_functions/3-print_all.c
@@ -34,7 +34,7 @@ void print_all(const char * const format, ...)
 				break;
 
 			case 'f':
-				f = va_arg(args, float);
+				f = va_arg(args, double);
 				printf("%f", f);
 				break;
